# backend/openapi.yaml
openapi: 3.0.3
info:
  title: StudyConnect Backend API
  version: 0.0.1
  description: |
    OpenAPI specification for the StudyConnect backend. Endpoints are protected by JWT bearer
    authentication (Keycloak) in production; dev profile allows unauthenticated access.
servers:
  - url: http://localhost:8080
    description: Local development server
security:
  - bearerAuth: []
tags:
  - name: Users
  - name: Groups
  - name: Tasks
paths:
  /api/users:
    get:
      tags: [Users]
      summary: List users
      responses:
        '200':
          description: Users fetched
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Users]
      summary: Create user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/users/{userId}:
    parameters:
      - in: path
        name: userId
        required: true
        schema:
          type: integer
          format: int64
    get:
      tags: [Users]
      summary: Get user by id
      responses:
        '200':
          description: User fetched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Users]
      summary: Update user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/users/me:
    get:
      tags: [Users]
      summary: Get current user from access token
      responses:
        '200':
          description: Current user fetched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/groups:
    get:
      tags: [Groups]
      summary: List public groups
      responses:
        '200':
          description: Groups fetched
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GroupResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Groups]
      summary: Create group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGroupRequest'
      responses:
        '201':
          description: Group created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/groups/search:
    parameters:
      - in: query
        name: query
        required: true
        schema:
          type: string
          minLength: 1
    get:
      tags: [Groups]
      summary: Search public groups by name
      responses:
        '200':
          description: Groups fetched
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GroupResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/groups/{groupId}:
    parameters:
      - in: path
        name: groupId
        required: true
        schema:
          type: integer
          format: int64
    get:
      tags: [Groups]
      summary: Get group by id
      responses:
        '200':
          description: Group fetched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Groups]
      summary: Update group (add members/admins)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGroupRequest'
      responses:
        '200':
          description: Group updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Groups]
      summary: Delete group
      responses:
        '204':
          description: Group deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/groups/membership/{userId}:
    parameters:
      - in: path
        name: userId
        required: true
        schema:
          type: integer
          format: int64
    get:
      tags: [Groups]
      summary: List groups a user belongs to
      responses:
        '200':
          description: Groups fetched
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GroupResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/groups/{groupId}/members:
    parameters:
      - in: path
        name: groupId
        required: true
        schema:
          type: integer
          format: int64
    get:
      tags: [Groups]
      summary: List members in a group
      responses:
        '200':
          description: Members fetched
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/groups/{groupId}/admins:
    parameters:
      - in: path
        name: groupId
        required: true
        schema:
          type: integer
          format: int64
    get:
      tags: [Groups]
      summary: List admins in a group
      responses:
        '200':
          description: Admins fetched
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/groups/{groupId}/members/{userId}:
    parameters:
      - in: path
        name: groupId
        required: true
        schema:
          type: integer
          format: int64
      - in: path
        name: userId
        required: true
        schema:
          type: integer
          format: int64
    delete:
      tags: [Groups]
      summary: Remove member from group
      responses:
        '204':
          description: Member removed
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/groups/{groupId}/admins/{userId}:
    parameters:
      - in: path
        name: groupId
        required: true
        schema:
          type: integer
          format: int64
      - in: path
        name: userId
        required: true
        schema:
          type: integer
          format: int64
    delete:
      tags: [Groups]
      summary: Remove admin from group
      responses:
        '204':
          description: Admin removed
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/groups/{groupId}/join:
    parameters:
      - in: path
        name: groupId
        required: true
        schema:
          type: integer
          format: int64
      - in: query
        name: userId
        required: true
        schema:
          type: integer
          format: int64
    post:
      tags: [Groups]
      summary: Join group by id
      responses:
        '204':
          description: Joined group
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/groups/join/{inviteCode}/{userId}:
    parameters:
      - in: path
        name: inviteCode
        required: true
        schema:
          type: string
      - in: path
        name: userId
        required: true
        schema:
          type: integer
          format: int64
    post:
      tags: [Groups]
      summary: Join group by invitation code
      responses:
        '204':
          description: Joined group
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/tasks/groups/{groupId}:
    parameters:
      - in: path
        name: groupId
        required: true
        schema:
          type: integer
          format: int64
    get:
      tags: [Tasks]
      summary: List tasks in a group
      responses:
        '200':
          description: Tasks fetched
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaskResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [Tasks]
      summary: Create task in group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskRequest'
      responses:
        '201':
          description: Task created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/tasks/users/{userId}:
    parameters:
      - in: path
        name: userId
        required: true
        schema:
          type: integer
          format: int64
    get:
      tags: [Tasks]
      summary: List tasks assigned to a user
      responses:
        '200':
          description: Tasks fetched
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaskResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/tasks/{taskId}:
    parameters:
      - in: path
        name: taskId
        required: true
        schema:
          type: integer
          format: int64
    put:
      tags: [Tasks]
      summary: Update task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskRequest'
      responses:
        '200':
          description: Task updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Tasks]
      summary: Delete task
      responses:
        '204':
          description: Task deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Provide a Keycloak-issued access token. Production profile requires the role configured as
        `required.keycloak.role` (default `studyconnect`).
  responses:
    BadRequest:
      description: Validation or business rule error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Missing required role/authority
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
          format: email
        lastname:
          type: string
        firstname:
          type: string
      required: [id, email, lastname, firstname]
    UserCreateRequest:
      type: object
      properties:
        firstname:
          type: string
          maxLength: 100
        lastname:
          type: string
          maxLength: 100
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 256
          description: Must include uppercase, lowercase, digit, and special character.
        email:
          type: string
          format: email
      required: [firstname, lastname, password, email]
    UserUpdateRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        firstname:
          type: string
          maxLength: 100
        lastname:
          type: string
          maxLength: 100
      required: [email, firstname, lastname]
    GroupResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        is_public:
          type: boolean
        created_by_id:
          type: integer
          format: int64
        created_at:
          type: string
          format: date-time
        last_updated_at:
          type: string
          format: date-time
        member_count:
          type: integer
        max_members:
          type: integer
        invite_code:
          type: string
      required:
        - id
        - name
        - is_public
        - created_by_id
        - created_at
        - last_updated_at
        - member_count
        - max_members
        - invite_code
    CreateGroupRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
        description:
          type: string
          maxLength: 500
        is_public:
          type: boolean
        max_members:
          type: integer
        member_ids:
          type: array
          items:
            type: integer
            format: int64
        admin_ids:
          type: array
          items:
            type: integer
            format: int64
        created_by_id:
          type: integer
          format: int64
      required:
        - name
        - is_public
        - max_members
        - member_ids
        - admin_ids
        - created_by_id
    UpdateGroupRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
        description:
          type: string
          maxLength: 500
        is_public:
          type: boolean
        member_ids:
          type: array
          items:
            type: integer
            format: int64
        admin_ids:
          type: array
          items:
            type: integer
            format: int64
    TaskResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        due_date:
          type: string
          format: date-time
        priority:
          $ref: '#/components/schemas/TaskPriority'
        status:
          $ref: '#/components/schemas/TaskStatus'
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        created_by_id:
          type: integer
          format: int64
        assignee_ids:
          type: array
          items:
            type: integer
            format: int64
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        last_status_change_at:
          type: string
          format: date-time
        group_id:
          type: integer
          format: int64
      required:
        - id
        - title
        - priority
        - status
        - created_by_id
        - assignee_ids
        - created_at
        - updated_at
        - group_id
    TaskRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 200
        description:
          type: string
          maxLength: 1000
        due_date:
          type: string
          format: date-time
        priority:
          $ref: '#/components/schemas/TaskPriority'
        status:
          $ref: '#/components/schemas/TaskStatus'
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        created_by_id:
          type: integer
          format: int64
        assignee_ids:
          type: array
          items:
            type: integer
            format: int64
      required:
        - title
        - priority
        - created_by_id
        - assignee_ids
    TaskPriority:
      type: string
      enum: [LOW, MEDIUM, HIGH]
    TaskStatus:
      type: string
      enum: [OPEN, IN_PROGRESS, COMPLETED, CANCELLED]
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string